fix double credit

+++ src/SmartCalc/credit_calc.c
@@ -1,18 +1,18 @@
 #include "credit_calc.h"
 
 // Аннуитетный ежемесячный платеж
-double A_MonthlyPayment(double credit_sum, int period, double percent) {
+double A_MonthlyPayment(double credit_sum, double period, double percent) {
   double monthly = 0.0;
   double month_percent =
-      (percent / 12 / 100);  // Рассчитываем процентную ставку в месяц
-  monthly = (credit_sum * month_percent * pow((1 + month_percent), period)) /
-            ((pow((1 + month_percent), period) - 1));
+      (percent / 12.0 / 100.0);  // Рассчитываем процентную ставку в месяц
+  monthly = (credit_sum * month_percent * pow((1.0 + month_percent), period)) /
+            ((pow((1 + month_percent), period) - 1.0));
 
   return monthly;
 }
 
 // Переплата при аннуитетном платеже
-double A_Overpayment(double credit_sum, int period, double monthly) {
+double A_Overpayment(double credit_sum, double period, double monthly) {
   double overpay = 0.0;
   overpay = (monthly * period) - credit_sum;  // Вычисляем переплату по формуле
   return overpay;
@@ -27,10 +27,10 @@ double TotalPayout(double credit_sum, double overpay) {
 }
 
 // Дифференцированный ежемесячный платеж
-double D_MonthlyPayment(double credit_sum, int period, double percent) {
+double D_MonthlyPayment(double credit_sum, double period, double percent) {
   double monthly = 0.0;
   double month_percent =
-      (percent / 12 / 100);  // периодическая процентная ставка
+      (percent / 12.0 / 100.0);  // периодическая процентная ставка
 
   monthly = (credit_sum / period) + (credit_sum * month_percent);
 
@@ -38,14 +38,14 @@ double D_MonthlyPayment(double credit_sum, int period, double percent) {
 }
 
 // Переплата при дифференцированном платеже
-double D_Overpayment(double credit_sum, int period, double percent) {
+double D_Overpayment(double credit_sum, double period, double percent) {
   double overpayment = 0.0;
   double body_payment =
       (credit_sum / period);  // сумма погашения тела кредита каждый месяц
   double remaining_debt = credit_sum;  // оставшаяся сумма долга
 
-  for (int i = 0; i < period; i++) {
-    double interest_payment = remaining_debt * (percent / 12 / 100);
+  for (double i = 0; i < period; i++) {
+    double interest_payment = remaining_debt * (percent / 12.0 / 100.0);
     remaining_debt -= body_payment;  // уменьшаем оставшуюся сумму долга на
                                      // сумму погашения тела кредита
     overpayment +=
@@ -56,7 +56,7 @@ double D_Overpayment(double credit_sum, int period, double percent) {
 }
 
 // Выбор типа ежемесячных платежей
-// double payment_type(char* type_pay, double credit_sum, int period,
+// double payment_type(char* type_pay, double credit_sum, double period,
 //                   double percent) {
 //     double result = 0;
 
@@ -88,11 +88,11 @@ double D_Overpayment(double credit_sum, int period, double percent) {
 //     return result;
 // }
 
-// int main() {
+// double main() {
 //  double credit_sum = 0.0;
 //  scanf("%lf", &credit_sum);
 
-//  int period = 0;
+//  double period = 0;
 //  scanf("%d", &period);
 
 //  double percent = 0.0;
